// ui/lib/api/schemas/engineCatalog.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 */
import type { EngineCatalogEntry } from './engineCatalogEntry'

export interface EngineCatalog {
  bubbleSegmenters: EngineCatalogEntry[]
  detectors: EngineCatalogEntry[]
  fontDetectors: EngineCatalogEntry[]
  inpainters: EngineCatalogEntry[]
  ocr: EngineCatalogEntry[]
  renderers: EngineCatalogEntry[]
  segmenters: EngineCatalogEntry[]
  translators: EngineCatalogEntry[]
}